refactor: updated ProviderWrapper location
