+++ __tests__/AuthScreen-test.tsx
@@ -6,7 +6,7 @@ import { Provider } from 'react-redux';
 import { configureStore } from '@reduxjs/toolkit';
 import { Linking } from 'react-native';
 import AuthApis from '../src/constants/apiConstant/AuthApi';
-import ProviderWrapper from './ProviderWrapper';
+import ProviderWrapper from '../src/utils/tests/ProviderWrapper';
 
 jest.mock('react-redux', () => {
   return {
@@ -25,7 +25,7 @@ it('AuthScreen is rendered', () => {
   render(
     <ProviderWrapper>
       <AuthScreen />
-    </ProviderWrapper>
+    </ProviderWrapper>,
   );
   screen.getByText(/welcome to/i);
   screen.getByText(/real dev squad/i);
@@ -49,7 +49,7 @@ it('Clicking on Sign in with Github opens browser', async () => {
   render(
     <ProviderWrapper>
       <AuthScreen />
-    </ProviderWrapper>
+    </ProviderWrapper>,
   );
 
   const githubSignInBtn = screen.getByText(Strings.SIGN_IN_BUTTON_TEXT);
